Run `terraform init` on each diff
We're seeing "uninitialized" errors when running this on Kubernetes with
git-sync, despite the fact that we're running `terraform init` at the start of
the process.

The cause is not well understood. We don't have logs (and we don't know why we
don't have logs!). However, on reflection, running `init` before each diff is
the only way to be sure that `init` has been run.